countdown template-package order payment page-user payment notf date-auth failure message

+++ src/main/java/com/thelastcodebenders/follower/dto/CountDownDTO.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class CountDownDTO {
-    private int endtimeSeconds;
+    private long endtimeSeconds;
 }
